fix: 24 real test failures — first full test run surfaced logic bugs
Compile is fully green; the test RUN now executes and failed on real
logic bugs (24 across common + agent). Fixes:

Real code bugs:
- corroborate.rs: Confidence enum ordered High<Medium<Low (derived
  Ord follows declaration) — reorder None<Low<Medium<High.
- wpa3.rs: WPA3-Enterprise (AKM 00:0f:ac:05) classified as
  Wpa2Enterprise because encryption_from_akm required SAE alongside —
  ac05 now maps directly to Wpa3Enterprise.
- wps.rs: recovered PIN kept the tool's surrounding quotes — trim.
- wep.rs: parse_iv_count now falls back to the first numeric token
  when the line has no '=' (status lines like 'Read 12345 packets').
- hid.rs: mouse delta sign-extension cast through i8 (0xFF → -1),
  matching the documented intent; u8→i16 gave +15.
- ble.rs: FitnessTracker classifier compared uppercase '180D'
  against lowercase {:04x} UUIDs — compare case-insensitively.
- evil_twin.rs: hostapd/dnsmasq config writers never created the run
  dir — create_dir_all first.
- scheduler.rs: run_pool joined workers unconditionally; a worker
  stuck in attack_fn stalled the caller 30s past the deadline. Now
  polls is_finished until the deadline and detaches stragglers.
- autopwn.rs: rank_targets fabricated wps_advertised=true from the
  encryption class alone (no WPS-IE evidence in the scan record) —
  stop guessing; explicit callers pass real beacon data.
- wps_crypto.rs: unify on the raw-digit search domain — checksum now
  subtracts b'0', build_full_pin maps bytes directly; docs updated.

Test-side fixes (wrong expectations/payloads):
- evil_twin: 'correcthorsebatterystaple' is 25 chars, not 27.
- hidden_beacon: IE/caps offsets ignored the 24-byte radiotap prefix.
- pmkid sanitize: non-ASCII BYTES each map to '_' (UTF-8 ë = 2 bytes).
- ble: AD length bytes wrong (len covers type+data): 0x03→0x04 for
  3-byte names, 0x05→0x04 for Eddystone; iBeacon payload built as a
  single proper 21-byte manufacturer struct with the real header.
- cracker: RECOVER-PLAINTEXT line must use the key.value contract.
- wep/wps_crypto/autopwn tests: contract corrections per the above.

Also: dead-code allows for corroborate/wpa3/report (scheduler-driven
modules like wps/wep), and evil_twin doc list reflows onto the item
line (rustdoc indent lint).

+++ crates/common/src/hid.rs
@@ -234,8 +234,10 @@ fn decode_generic_mouse(payload: &[u8]) -> HidReport {
         return HidReport::Malformed;
     }
     let buttons = payload[0];
-    let dx = i16::from(payload[1]) >> 4; // sign-extend high nibble
-    let dy = i16::from(payload[2]) >> 4;
+    // Sign-extend: reinterpret each byte as i8 first, THEN widen and shift,
+    // so 0xFF decodes to -1 (right move) not +15.
+    let dx = (payload[1] as i8 as i16) >> 4;
+    let dy = (payload[2] as i8 as i16) >> 4;
     HidReport::Mouse {
         dx,
         dy,
